Extracted method and version to traits
This should allow our application to maintain all the related functions
in lean and easy to read components that the request method just brings
together.

This should make maintaining and debugging the small components
significantly easier.
